fix bugs and tests

+++ src/main/java/ru/yandex/practicum/filmorate/controller/FilmController.java
@@ -3,6 +3,7 @@ package ru.yandex.practicum.filmorate.controller;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.web.bind.annotation.*;
+import ru.yandex.practicum.filmorate.exception.IncorrectParameterException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.service.FilmService;
 
@@ -29,7 +30,7 @@ public class FilmController {
         return filmService.findFilmById(filmId);
     }
 
-    @GetMapping("/films/popular")
+    @GetMapping("/popular")
     public List<Film> findMostPopularFilms(
             @RequestParam(value = "count", required = false, defaultValue = "10") int count
     ) {
@@ -39,17 +40,20 @@ public class FilmController {
     @PostMapping
     public Film create(@RequestBody Film film) {
         log.debug("Film: {} added", film.getName());
-        return film;
+        return filmService.createFilm(film);
     }
 
     @PutMapping
     public Film update(@RequestBody Film film) {
         log.debug("Film: {} updated", film.getName());
-        return film;
+        return filmService.updateFilm(film);
     }
 
     @PutMapping("/{id}/like/{userId}")
     public void addLike(@PathVariable("id") int filmId, @PathVariable("userId") int userId) {
+        if (userId < 0) {
+            throw new IncorrectParameterException("userId");
+        }
         log.debug("Like from user with id: {} to film: {} added",
                 userId, filmService.findFilmById(filmId).getName());
         filmService.addLike(filmId, userId);
@@ -57,6 +61,9 @@ public class FilmController {
 
     @DeleteMapping("/{id}/like/{userId}")
     public void deleteLike(@PathVariable("id") int filmId, @PathVariable("userId") int userId) {
+        if (userId < 0) {
+            throw new IncorrectParameterException("userId");
+        }
         log.debug("Like from user with id: {} to film: {} deleted",
                 userId, filmService.findFilmById(filmId).getName());
         filmService.deleteLike(filmId, userId);
